neural-nexus-backend/requirements.txt
aider-chat==0.62.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
defusedxml==0.7.1
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
Django==5.0
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
einops==0.8.0
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.6
httpie==3.2.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
itypes==1.2.0
Jinja2==3.1.4
jiter==0.7.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
kornia==0.7.3
kornia_rs==0.1.5
litellm==1.51.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
openai==1.53.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pypandoc==1.14
pyperclip==1.9.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.1
safetensors==0.4.5
scipy==1.13.1
sentencepiece==0.2.0
setuptools==75.0.0
setuptools-scm==8.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
spandrel==0.3.4
sqlparse==0.5.0
sympy==1.13.2
tiktoken==0.8.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchsde==0.2.6
torchvision==0.19.1
tqdm==4.66.6
trampoline==0.1.2
transformers==4.44.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
XlsxWriter==3.2.0
yarl==1.17.1
zipp==3.20.2
